@@ -1,5 +1,5 @@
 import React from 'react';
-import Loadable from 'react-loadable';
+import Loadable from '@yeutech-lab/react-loadable';
 import Loading from './Loading';
 import delay from '../utils/delay';
 import path from 'path';
@@ -10,5 +10,5 @@ const LoadableExample = Loadable({
 });
 
 export default function App() {
-  return <LoadableExample/>;
+  return <LoadableExample />;
 }
