@@ -1,5 +1,5 @@
 import React from 'react';
-import Loadable from 'react-loadable';
+import Loadable from '@yeutech-lab/react-loadable';
 import Loading from './Loading';
 import delay from '../utils/delay';
 import path from 'path';
@@ -13,7 +13,7 @@ export default function Example() {
   return (
     <div>
       <h1>Hello from a loadable component</h1>
-      <LoadableNested/>
+      <LoadableNested />
     </div>
   );
 }
