@@ -3,7 +3,7 @@ export default function({ types: t, template }) {
     visitor: {
       ImportDeclaration(path) {
         let source = path.node.source.value;
-        if (source !== 'react-loadable') return;
+        if (source !== '@yeutech-lab/react-loadable') return;
 
         let defaultSpecifier = path.get('specifiers').find(specifier => {
           return specifier.isImportDefaultSpecifier();
